dense layers refactored
